OEL-100: Reuse language code. Ask for validation in TODO.
